вынес функцию render в модуль util
