+++ src/main.js
@@ -5,30 +5,10 @@ import {createSortTemplate} from "./View/sorting.js";
 import {createTaskAddTemplate} from "./View/task-card-add.js";
 import {createTaskCardTemplate} from "./View/task-card.js";
 import {createLoadButton} from "./View/load-button.js";
-
-const IterationCount = {
-  DEFAULT: 1,
-  CARD: 5
-};
-
-const InsertPlace = {
-  BEFOREEND: `beforeend`,
-  AFTERBEGIN: `afterbegin`
-};
+import {IterationCount, InsertPlace, render} from "./util.js";
 
 const siteMainElement = document.querySelector(`.main`);
 const siteHeaderElement = siteMainElement.querySelector(`.main__control`);
-
-const render = ({
-  container,
-  template,
-  place = InsertPlace.BEFOREEND,
-  iteration = IterationCount.DEFAULT
-}) => {
-  for (let i = 0; i < iteration; i++) {
-    container.insertAdjacentHTML(place, template);
-  }
-};
 
 render({container: siteHeaderElement, template: createSiteMenuTemplate()});
 render({container: siteMainElement, template: createFilterTemplate()});
