+++ src/util.js
@@ -1,0 +1,22 @@
+const IterationCount = {
+  DEFAULT: 1,
+  CARD: 5
+};
+
+const InsertPlace = {
+  BEFOREEND: `beforeend`,
+  AFTERBEGIN: `afterbegin`
+};
+
+const render = ({
+  container,
+  template,
+  place = InsertPlace.BEFOREEND,
+  iteration = IterationCount.DEFAULT
+}) => {
+  for (let i = 0; i < iteration; i++) {
+    container.insertAdjacentHTML(place, template);
+  }
+};
+
+export {IterationCount, InsertPlace, render};
